+++ algosdk/encoding.py
@@ -1,4 +1,5 @@
 import base64
+import warnings
 from collections import OrderedDict
 from typing import Union
 
@@ -97,6 +98,9 @@ def future_msgpack_decode(enc):
 
 def msgpack_decode(enc):
     """
+    NOTE: This method is deprecated:
+    Please use `future_msgpack_decode` instead.
+
     Decode a msgpack encoded object from a string.
 
     Args:
@@ -106,6 +110,11 @@ def msgpack_decode(enc):
         Transaction, SignedTransaction, Multisig, Bid, or SignedBid:\
             decoded object
     """
+    warnings.warn(
+        "`msgpack_decode` is being deprecated. "
+        "Please use `future_msgpack_decode` instead.",
+        DeprecationWarning,
+    )
     decoded = enc
     if not isinstance(enc, dict):
         decoded = msgpack.unpackb(base64.b64decode(enc), raw=False)