+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - `class StateSchema`'s method `undictify()` now returns a `StateSchema` object instead of a python `dict`
 
+# v1.18.0
+### Enhancements
+* Deprecation: Add deprecation warnings on v1 algod API and old transaction format by @algochoi in https://github.com/algorand/py-algorand-sdk/pull/381
+* enhancement: add unit test for ParticipationUpdates field by @shiqizng in https://github.com/algorand/py-algorand-sdk/pull/386
+
 # v1.17.0
 ## What's Changed
 ### Bugfixes